socialmedia/requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bcrypt==3.2.0
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.12.5
cffi==1.14.4
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
Django==3.1.7
django-active-link==0.1.7
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
Faker==6.0.0
ipykernel @ file:///C:/ci/ipykernel_1607454116140/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1612212932352/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1606932640618/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213515523/work
misaka==2.1.1
mysqlclient==2.0.3
parso==0.7.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.1.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycparser==2.20
Pygments @ file:///tmp/build/80754af9/pygments_1613849567472/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==228
pyzmq==20.0.0
six @ file:///C:/ci/six_1605205404068/work
soupsieve==2.2
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
tornado @ file:///C:/ci/tornado_1606924294691/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
wincertstore==0.2
